requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
httplib2==0.15.0
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pickleshare==0.7.5
plotly==4.8.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
spyder-kernels==1.9.1
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.3
webencodings==0.5.1
wincertstore==0.2
zipp==3.1.0
